+++ packages/eslint-plugin/rules/no-i18n-in-save.js
@@ -25,6 +25,8 @@ module.exports = {
 		const normalizedFilename = filename.replace( /\\/g, '/' );
 		const isDeprecatedFile =
 			normalizedFilename.includes( '/deprecated.js' ) ||
+			normalizedFilename.includes( '/deprecated.mjs' ) ||
+			normalizedFilename.includes( '/deprecated.cjs' ) ||
 			normalizedFilename.includes( '/deprecated.ts' ) ||
 			normalizedFilename.includes( '/deprecated.mts' ) ||
 			normalizedFilename.includes( '/deprecated.cts' ) ||
@@ -37,6 +39,8 @@ module.exports = {
 
 		const isSaveFile =
 			normalizedFilename.endsWith( '/save.js' ) ||
+			normalizedFilename.endsWith( '/save.mjs' ) ||
+			normalizedFilename.endsWith( '/save.cjs' ) ||
 			normalizedFilename.endsWith( '/save.ts' ) ||
 			normalizedFilename.endsWith( '/save.mts' ) ||
 			normalizedFilename.endsWith( '/save.cts' ) ||